+++ WarehouseAPI/Data/WarehouseContextSeed.cs
@@ -40,7 +40,7 @@ namespace WarehouseAPI.Data
                     Supplier = "ABC LTD",
                     Quantity = quantity,
                     Price = product.Price,
-                    Discount = Math.Round(rand.NextDouble() * 0.2, 2),
+                    Discount = (decimal)Math.Round(rand.NextDouble() * 0.2, 2),
                     Sold = sold,
                     AvailableOnStock = availableOnStock
                 });
