+++ WebUI/Pages/CheckOut.cshtml.cs
@@ -123,7 +123,7 @@ namespace WebUI.Pages
                     await _customerService.UpdateCustomer(customer);
                 }
 
-                Order.TotalPrice = (decimal)ShoppingCart.TotalPrice;
+                Order.TotalPrice = ShoppingCart.TotalPrice;
 
                 foreach (var item in ShoppingCart.Items)
                 {
